+++ Editor/Scripts/Post Process Build/ManageViewsPostProcessBuild.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+using UnityEditor;
+using UnityEditor.Callbacks;
+
+public class ManageViewsPostProcessBuild
+{
+    [PostProcessBuildAttribute(0)]
+    public static void OnPostprocessBuild(BuildTarget target, string pathToBuiltProject)
+    {
+        //PlayModeStateChanged.ResetViewsAfterPlay();
+    }
+
+}